# api/openapi.yaml
openapi: 3.0.3
info:
  title: Opero API
  version: 0.1.0
  description: >
    Opero people-operations API. This spec is the single source of truth —
    every operation is defined here before it is implemented. Operational
    routes like /health are infrastructure and intentionally not part of the
    contract.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /departments:
    get:
      operationId: listDepartments
      tags: [identity]
      summary: List departments.
      description: Returns all departments for the tenant. Org hierarchy is expressed via each department's nullable parent_id.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The tenant's departments.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Department'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createDepartment
      tags: [identity]
      summary: Create a department.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDepartmentRequest'
      responses:
        '201':
          description: Department created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /departments/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      operationId: getDepartment
      tags: [identity]
      summary: Get a department by id.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The department.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updateDepartment
      tags: [identity]
      summary: Update a department.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDepartmentRequest'
      responses:
        '200':
          description: Updated department.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteDepartment
      tags: [identity]
      summary: Delete a department.
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /employees:
    get:
      operationId: listEmployees
      tags: [identity]
      summary: List employees.
      description: Returns the tenant's employees, optionally filtered by department or status.
      security:
        - bearerAuth: []
      parameters:
        - name: department_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, inactive]
      responses:
        '200':
          description: The tenant's employees.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Employee'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createEmployee
      tags: [identity]
      summary: Create an employee.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmployeeRequest'
      responses:
        '201':
          description: Employee created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /employees/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      operationId: getEmployee
      tags: [identity]
      summary: Get an employee by id.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The employee.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updateEmployee
      tags: [identity]
      summary: Update an employee.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEmployeeRequest'
      responses:
        '200':
          description: Updated employee.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteEmployee
      tags: [identity]
      summary: Delete an employee.
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /roles:
    get:
      operationId: listRoles
      tags: [identity]
      summary: List roles.
      description: Returns the tenant's roles. Each role carries a set of permission keys (not yet enforced in v1).
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The tenant's roles.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createRole
      tags: [identity]
      summary: Create a role.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest'
      responses:
        '201':
          description: Role created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: A role with this name already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /roles/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      operationId: getRole
      tags: [identity]
      summary: Get a role by id.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updateRole
      tags: [identity]
      summary: Update a role.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
      responses:
        '200':
          description: Updated role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteRole
      tags: [identity]
      summary: Delete a role.
      description: Deletes a role. Employees assigned to it have their role cleared (set to null).
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /employees/{id}/login:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    post:
      operationId: createEmployeeLogin
      tags: [identity]
      summary: Provision a login for an employee.
      description: >
        Creates a control-plane user account for the employee and links it
        (employees.user_id), so the employee can authenticate (e.g. field staff
        on the mobile app). Fails if the employee already has a login or the
        email is already taken in this tenant.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLoginRequest'
      responses:
        '201':
          description: Login created and linked to the employee.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: The employee already has a login, or the email is taken.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication failed or is missing.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: A stable, machine-readable error code.
          example: invalid_credentials
        message:
          type: string
          description: A human-readable explanation.
...
    UserSummary:
      type: object
      required: [id, email, role, status]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, manager, employee]
        status:
          type: string
          enum: [active, disabled]
...
    Department:
      type: object
      required: [id, name, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        parent_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateDepartmentRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        parent_id:
          type: string
          format: uuid
          nullable: true
    UpdateDepartmentRequest:
      type: object
      description: >
        PATCH semantics — only provided fields are changed. Fields cannot be
        cleared back to null via PATCH in v1 (omitting a field and sending null
        are equivalent).
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        parent_id:
          type: string
          format: uuid
          nullable: true
    Role:
      type: object
      required: [id, name, permissions, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        permissions:
          type: array
          description: Permission keys granted by this role. Stored but not enforced in v1.
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateRoleRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          nullable: true
        permissions:
          type: array
          items:
            type: string
    UpdateRoleRequest:
      type: object
      description: >
        PATCH semantics — only provided fields are changed. permissions, when
        provided, replaces the whole set.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          nullable: true
        permissions:
          type: array
          items:
            type: string
    Employee:
      type: object
      required: [id, full_name, employment_type, status, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
          nullable: true
          description: Links to the control-plane users.id (different database; no FK).
        role_id:
          type: string
          format: uuid
          nullable: true
          description: The employee's assigned role (access-control), if any.
        full_name:
          type: string
        email:
          type: string
          format: email
          nullable: true
        phone:
          type: string
          nullable: true
        employment_type:
          type: string
          enum: [full_time, part_time, freelance, seasonal]
        department_id:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
        hired_at:
          type: string
          format: date
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateEmployeeRequest:
      type: object
      required: [full_name, employment_type]
      properties:
        full_name:
          type: string
          minLength: 1
          maxLength: 200
        employment_type:
          type: string
          enum: [full_time, part_time, freelance, seasonal]
        email:
          type: string
          format: email
          nullable: true
        phone:
          type: string
          nullable: true
        department_id:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
          nullable: true
        hired_at:
          type: string
          format: date
          nullable: true
        user_id:
          type: string
          format: uuid
          nullable: true
        role_id:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
          enum: [active, inactive]
          description: Optional on create; defaults to active.
    UpdateEmployeeRequest:
      type: object
      description: >
        PATCH semantics — only provided fields are changed. Fields cannot be
        cleared back to null via PATCH in v1 (omitting a field and sending null
        are equivalent).
      properties:
        full_name:
          type: string
          minLength: 1
          maxLength: 200
        employment_type:
          type: string
          enum: [full_time, part_time, freelance, seasonal]
        email:
          type: string
          format: email
          nullable: true
        phone:
          type: string
          nullable: true
        department_id:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
          nullable: true
        hired_at:
          type: string
          format: date
          nullable: true
        role_id:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
          enum: [active, inactive]
...
    CreateLoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 320
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
        role:
          type: string
          enum: [admin, manager, employee]
          description: Defaults to employee when omitted.
...